mm: zram: use rw_page for zram page but submit_bio for zram wb page
Since reading zram wb page is an async operation, we can not immediately
free the slot here. So we need to read the page via submit_bio and free
the slot in end_swap_bio_read. (bio chain will be used to handle bio here
and bio in zram driver for wb page).

Change-Id: I71b88c4a2388c086bd7e0cd63b8c632b93217a95
Signed-off-by: huangzq2 <huangzq2@motorola.com>
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>